+++ packages/atproto/lib/src/repositories/repositories_service.dart
@@ -65,8 +65,6 @@ abstract class RepositoriesService {
   ///
   /// ## Parameters
   ///
-  /// - [did]: The DID of repo.
-  ///
   /// - [uri]: The AT URI of record.
   ///
   /// - [cid]: The CID of the version of the record. If not specified,
@@ -80,7 +78,6 @@ abstract class RepositoriesService {
   ///
   /// - https://github.com/bluesky-social/atproto/blob/main/lexicons/com/atproto/repo/getRecord.json
   Future<core.XRPCResponse<RecordValue>> findRecord({
-    required String did,
     required core.AtUri uri,
     String? cid,
   });
@@ -207,14 +204,13 @@ class _RepositoriesService extends ATProtoBaseService
 
   @override
   Future<core.XRPCResponse<RecordValue>> findRecord({
-    required String did,
     required core.AtUri uri,
     String? cid,
   }) async =>
       await super.get(
         'getRecord',
         parameters: {
-          'repo': did,
+          'repo': uri.hostname,
           'collection': uri.collection,
           'rkey': uri.rkey,
           'cid': cid,